Rename Erc20ApiImpl
