+++ src/api/erc20/Erc20ApiMock.ts
@@ -5,7 +5,7 @@ import { ZERO, ALLOWANCE_MAX_VALUE } from 'const'
 import { RECEIPT } from '../../../test/data'
 import { log, assert } from 'utils'
 import { waitAndSendReceipt } from 'utils/mock'
-import { Erc20Api } from './Erc20ApiImpl'
+import { Erc20Api } from './Erc20Api'
 
 interface Balances {
   [userAddress: string]: { [tokenAddress: string]: BN }
