+++ src/api/index.ts
@@ -4,7 +4,7 @@ import { WalletApiImpl, WalletApi } from './wallet/WalletApiImpl'
 import { TokenListApiImpl, TokenList } from './tokenList/TokenListApiImpl'
 import { TokenListApiMock } from './tokenList/TokenListApiMock'
 import { Erc20ApiMock } from './erc20/Erc20ApiMock'
-import { Erc20ApiImpl, Erc20Api } from './erc20/Erc20ApiImpl'
+import { Erc20ApiImpl, Erc20Api } from './erc20/Erc20Api'
 import { DepositApiMock } from './exchange/DepositApiMock'
 import { DepositApiImpl, DepositApi } from './exchange/DepositApiImpl'
 import { ExchangeApiImpl, ExchangeApi } from './exchange/ExchangeApiImpl'
