+++ src/services/factories/addTokenToExchange.ts
@@ -4,7 +4,7 @@ import { log, getImageUrl } from 'utils'
 
 import { getErc20Info } from '../helpers'
 import { ExchangeApi } from 'api/exchange/ExchangeApiImpl'
-import { Erc20Api } from 'api/erc20/Erc20ApiImpl'
+import { Erc20Api } from 'api/erc20/Erc20Api'
 
 interface Params {
   exchangeApi: ExchangeApi
