+++ src/services/helpers/getErc20Info.ts
@@ -3,7 +3,7 @@ import Web3 from 'web3'
 import { MinimalTokenDetails } from 'types'
 import { log } from 'utils'
 import { DEFAULT_PRECISION } from 'const'
-import { Erc20Api } from 'api/erc20/Erc20ApiImpl'
+import { Erc20Api } from 'api/erc20/Erc20Api'
 
 /**
  * Wraps erc20 function and returns undefined in case of failure
